Moving dequant util files to deepseek_v3_d_p
